+++ nexus/users/tests/test_models.py
@@ -18,7 +18,7 @@ class UserModelTestCase(TestCase):
         assert u.email == 'f@f.com'
         assert u.get_full_name() == 'F B'
         assert u.get_short_name() == 'F'
-        assert str(u) == str(u.id)
+        assert str(u) == str('{} - {}'.format(u.email, u.get_full_name()))
         assert u.gender == User.GENDER_CHOICES.OTHERS
         assert u.tshirt_size == User.TSHIRT_SIZE_CHOICES.LARGE
         assert u.is_core_organizer is False
@@ -30,4 +30,4 @@ class UserModelTestCase(TestCase):
         assert u.is_active is True
         assert u.is_staff is True
         assert u.is_superuser is True
-        assert str(u) == str(u.id)
+        assert str(u) == str('{} - {}'.format(u.email, u.get_full_name()))